+++ src/highlight.rs
@@ -107,7 +107,6 @@ fn highlight(source: &str, ext: &str) -> Result<String, Error> {
         };
 
         line_number += 1;
-        let formatted_str = formatted.as_str();
         let line_number = format!(
             r#"<tr><td class="line-number" id="L{line_number}"><a href=#L{line_number}>{line_number:>4}</a></td>"#
         );
@@ -118,7 +117,10 @@ fn highlight(source: &str, ext: &str) -> Result<String, Error> {
             html.push_str(&"<span>".repeat(delta.abs().try_into()?));
         }
 
-        html.push_str(&formatted_str);
+        // Strip stray newlines that cause vertically stretched lines.
+        for c in formatted.chars().filter(|c| *c != '\n') {
+            html.push(c);
+        }
 
         if delta > 0 {
             html.push_str(&"</span>".repeat(delta.try_into()?));